Model-implied moments for a penfa object

An S4 method returning the model-implied moments for an object of class penfa. For every group, a list with the model-implied moments is returned: cov contains the implied covariance matrix, and mean the implied mean vector. If just the covariance matrix is analyzed, only the cov argument is returned.

# S4 method for penfa
fitted(object, labels = TRUE)

Arguments

object
An object of class penfa, found as a result of a call to penfa.

labels
Logical. If TRUE, the model-implied moments are named according to the item names used in the model syntax.

See also

Examples

data(ccdata)

syntax = 'help  =~   h1 + h2 + h3 + h4 + h5 + h6 + h7 + 0*v1 + v2 + v3 + v4 + v5
          voice =~ 0*h1 + h2 + h3 + h4 + h5 + h6 + h7 +   v1 + v2 + v3 + v4 + v5'

alasso_fit <- penfa(## factor model
                    model  = syntax,
                    data   = ccdata,
                    std.lv = TRUE,
                    ## penalization
                    pen.shrink = "alasso",
                    eta = list(shrink = c("lambda" = 0.01), diff = c("none" = 0)),
                    ## automatic procedure
                    strategy = "auto")

#> Computing weights for alasso (ML estimates)... done.
#> 
#> Automatic procedure: 
#> Iteration  1 : 0.00298271 
#> Iteration  2 : 0.00452604 
#> 
#> Largest absolute gradient value: 12.76355181
#> Fisher information matrix is positive definite
#> Eigenvalue range: [180.2917, 9189645]
#> Trust region iterations: 15 
#> Factor solution: admissible 
#> Effective degrees of freedom: 27.12936  

fitted(alasso_fit)

#> $cov
#>    h1    h2    h3    h4    h5    h6    h7    v1    v2    v3    v4    v5   
#> h1 0.974                                                                  
#> h2 0.657 0.969                                                            
#> h3 0.594 0.665 0.973                                                      
#> h4 0.678 0.759 0.686 0.968                                                
#> h5 0.655 0.734 0.664 0.757 0.968                                          
#> h6 0.669 0.749 0.677 0.772 0.749 0.966                                    
#> h7 0.630 0.706 0.638 0.724 0.708 0.727 0.967                              
#> v1 0.571 0.640 0.579 0.652 0.649 0.672 0.681 0.969                        
#> v2 0.585 0.656 0.593 0.668 0.665 0.689 0.697 0.741 0.968                  
#> v3 0.566 0.634 0.573 0.645 0.642 0.665 0.674 0.716 0.734 0.970            
#> v4 0.566 0.634 0.573 0.646 0.643 0.666 0.674 0.717 0.735 0.710 0.970      
#> v5 0.541 0.606 0.547 0.617 0.614 0.636 0.644 0.685 0.701 0.678 0.678 0.972
#>