penfa Options

The default options internally used by the penfa function. These options can be changed by passing "name = value" arguments to the penfa function call, where they are being added to the "..." argument.

penfaOptions(
  opt = list(meanstructure = FALSE, int.ov.free = FALSE, int.lv.free = FALSE,
    orthogonal = FALSE, std.lv = FALSE, auto.fix.first = FALSE, auto.fix.single = FALSE,
    std.ov = FALSE, information = "fisher", control = list(), optim.dx.tol = 100, a.scad
    = 3.7, a.mcp = 3, a.alasso = 1, weights = NULL, cbar = 1e-08, gamma = 4, user.start =
    FALSE, start.val = c(), verbose = TRUE, warn = TRUE, debug = FALSE)
)

Arguments

opt
List of default options. See below for details.

Value

A list of default options internally used by the penfa function.

Details

The following section details the full list of options currently accepted by the penfa function.

Model features:

meanstructure:

Logical. If TRUE, a meanstructure is requested. It should be used in conjunction with int.ov.free and int.lv.free or intercept-like formulas in the model syntax. Default to FALSE.

int.ov.free:

Logical. If FALSE, the intercepts of the observed variables are fixed to zero. Default to FALSE.

int.lv.free:

Logical. If FALSE, the intercepts of the common factors are fixed to zero. Default to FALSE.

orthogonal:

Logical. If TRUE, all covariances among the common factors are set to zero. Default to FALSE.

std.lv:

Logical. If TRUE, the factor variances are fixed to 1.0. Default to FALSE.

auto.fix.first:

Logical. If TRUE, the factor loading of the first indicator is set to 1.0 for every factor. Default to FALSE.

auto.fix.single:

Logical. If TRUE, the residual variance (if included) of an observed indicator is set to zero if it is the only indicator of a common factor. Default to FALSE.

Data options:

std.ov:

Logical. If TRUE, all observed variables are standardized before entering the analysis. Default to FALSE.

Estimation and optimization:

information:

Character. If "fisher", the penalized expected Fisher information matrix is used as second-order derivatives in the trust-region algorithm and for computing the standard errors of the model parameters. If "hessian", the penalized Hessian matrix is used. Default to "fisher".

control:

A list containing control parameters passed to the trust-region optimizer. See the manual page of trust from the trust package for an overview of its control parameters. Default values for these parameters are rinit=1L, rmax=100L, iterlim=1000L,
fterm = sqrt(.Machine$double.eps), mterm = sqrt(.Machine$double.eps).

optim.dx.tol

Numeric. The tolerance value used when checking the size of the elements of the gradient of the objective function. Default equal to 100.

Penalization:

a.scad

Numeric. The shape parameter for the scad penalty. Default to 3.7, as recommended by Fan & Li (2001).

a.mcp

Numeric. The shape parameter of the mcp penalty. Default to 3.

a.alasso

Numeric. The exponent in the adaptive weights for the alasso penalty. Default to 1.

weights

Numeric. Only valid when either pen.shrink or pen.diff is equal to "alasso". An optional vector of values provided by the user representing a consistent estimate for each model parameter. The vector is then internally used for computing the adaptive weights. If unspecified, the maximum likelihood estimates (MLE) from the unpenalized model are used.

cbar

Numeric. Numerical constant used in the local approximation of the penalty functions. Default to 1e-08.

Automatic procedure:
gamma

Numeric. The value of the influence factor used in the automatic tuning parameter procedure. Default to 4.

user.start

Logical whether the user has provided a vector of starting values for the model parameter estimates.

start.val

Numeric. An optional vector of parameter estimates to be used as starting values for the model parameters. This option is also internally used by the automatic procedure.

Verbosity options:

verbose:

Logical. If TRUE, some information on the estimation process (e.g., convergence and admissibility checks, effective degrees of freedom) are printed out. Default to TRUE.

warn:

Logical. If TRUE, some warnings are printed out during the iterations. Default to TRUE.

debug:

Logical. If TRUE, debugging information is printed out. Default to FALSE.