Print parameter estimates in table format

The parameter estimates of the penalized factor analysis model in each group.

penfaParEstim(
  object,
  se = TRUE,
  ci = TRUE,
  level = 0.95,
  remove.nonfree = FALSE,
  output = "data.frame",
  header = FALSE
)

Arguments

object
An object of class penfa.

se
Logical. If TRUE, it includes a column with the standard errors.

ci
Logical. If TRUE, the confidence intervals are added to the output.

level
The confidence level, default is 0.95.

remove.nonfree
Logical. If TRUE, it filters the output and removes all rows with fixed (that is, neither free, nor penalized) parameters.

output
Character. If "data.frame", the parameter table is displayed as a standard formatted data.frame. If "text", the parameter table is displayed with subsections (as used by the summary function).

header
Logical, only used if output = "text". If TRUE, it prints a header on top of the parameter list with details on the group levels and the information matrix used during optimization by the trust-region algorithm.

Value

A dataframe of class penfa.data.frame with the parameter estimates of a penfa model for each group.

See also

Examples

data(ccdata)

syntax = 'help  =~   h1 + h2 + h3 + h4 + h5 + h6 + h7 + 0*v1 + v2 + v3 + v4 + v5
          voice =~ 0*h1 + h2 + h3 + h4 + h5 + h6 + h7 +   v1 + v2 + v3 + v4 + v5'

alasso_fit <- penfa(## factor model
                    model  = syntax,
                    data   = ccdata,
                    std.lv = TRUE,
                    ## penalization
                    pen.shrink = "alasso",
                    eta = list(shrink = c("lambda" = 0.01), diff = c("none" = 0)),
                    ## automatic procedure
                    strategy = "auto")

#> Computing weights for alasso (ML estimates)... done.
#> 
#> Automatic procedure: 
#> Iteration  1 : 0.00298271 
#> Iteration  2 : 0.00452604 
#> 
#> Largest absolute gradient value: 12.76355181
#> Fisher information matrix is positive definite
#> Eigenvalue range: [180.2917, 9189645]
#> Trust region iterations: 15 
#> Factor solution: admissible 
#> Effective degrees of freedom: 27.12936  

penfaParEstim(alasso_fit)

#>      lhs op   rhs  type penalty           est           se      ci_0.025
#> 1   help =~    h1   pen  shrink  7.658121e-01 0.0300355828  0.7069434576
#> 2   help =~    h2   pen  shrink  8.579551e-01 0.0281200574  0.8028407715
#> 3   help =~    h3   pen  shrink  7.753391e-01 0.0298554161  0.7168235129
#> 4   help =~    h4   pen  shrink  9.208458e-01 0.0378462170  0.8466685372
#> 5   help =~    h5   pen  shrink  8.096618e-01 0.0395996245  0.7320479239
#> 6   help =~    h6   pen  shrink  7.820608e-01 0.0438313584  0.6961528687
#> 7   help =~    h7   pen  shrink  5.230518e-01 0.0495779689  0.4258807765
#> 8   help =~    v1 fixed    none  0.000000e+00 0.0000000000  0.0000000000
#> 9   help =~    v2   pen  shrink  4.013798e-05 0.0013109918 -0.0025293588
#> 10  help =~    v3   pen  shrink  2.389271e-05 0.0012510188 -0.0024280590
#> 11  help =~    v4   pen  shrink  8.563583e-07 0.0006059277 -0.0011867401
#> 12  help =~    v5   pen  shrink -9.202527e-07 0.0003299068 -0.0006475258
#> 13 voice =~    h1 fixed    none  0.000000e+00 0.0000000000  0.0000000000
#> 14 voice =~    h2   pen  shrink -1.287100e-05 0.0009545266 -0.0018837087
#> 15 voice =~    h3   pen  shrink  5.304850e-06 0.0006379165 -0.0012449884
#> 16 voice =~    h4   pen  shrink -4.119937e-02 0.0279540523 -0.0959883072
#> 17 voice =~    h5   pen  shrink  5.262033e-02 0.0311089159 -0.0083520210
#> 18 voice =~    h6   pen  shrink  1.042524e-01 0.0384810259  0.0288309311
#> 19 voice =~    h7   pen  shrink  3.412021e-01 0.0487496049  0.2456545891
#> 20 voice =~    v1   pen  shrink  8.507219e-01 0.0283398664  0.7951768003
#> 21 voice =~    v2   pen  shrink  8.714683e-01 0.0278826758  0.8168192712
#> 22 voice =~    v3   pen  shrink  8.418956e-01 0.0285565995  0.7859257263
#> 23 voice =~    v4   pen  shrink  8.429135e-01 0.0285202472  0.7870148371
#> 24 voice =~    v5   pen  shrink  8.047512e-01 0.0293422229  0.7472414679
#> 25    h1 ~~    h1  free    none  3.875064e-01 0.0213756233  0.3456109291
#> 26    h2 ~~    h2  free    none  2.331226e-01 0.0141372069  0.2054141415
#> 27    h3 ~~    h3  free    none  3.723067e-01 0.0206500165  0.3318334075
#> 28    h4 ~~    h4  free    none  1.844461e-01 0.0124028238  0.1601370438
#> 29    h5 ~~    h5  free    none  2.349729e-01 0.0140838417  0.2073691125
#> 30    h6 ~~    h6  free    none  2.009585e-01 0.0123597415  0.1767338671
#> 31    h7 ~~    h7  free    none  2.638821e-01 0.0147092960  0.2350523666
#> 32    v1 ~~    v1  free    none  2.454065e-01 0.0151161227  0.2157794305
#> 33    v2 ~~    v2  free    none  2.083625e-01 0.0135010799  0.1819008483
#> 34    v3 ~~    v3  free    none  2.607043e-01 0.0158005083  0.2297358369
#> 35    v4 ~~    v4  free    none  2.590440e-01 0.0157260085  0.2282215497
#> 36    v5 ~~    v5  free    none  3.240195e-01 0.0187007932  0.2873666333
#> 37  help ~~  help fixed    none  1.000000e+00 0.0000000000  1.0000000000
#> 38 voice ~~ voice fixed    none  1.000000e+00 0.0000000000  1.0000000000
#> 39  help ~~ voice  free    none  8.771535e-01 0.0114966949  0.8546204111
#>        ci_0.975
#> 1  0.8246807786
#> 2  0.9130693711
#> 3  0.8338545936
#> 4  0.9950229819
#> 5  0.8872755994
#> 6  0.8679686366
#> 7  0.6202228433
#> 8  0.0000000000
#> 9  0.0026096347
#> 10 0.0024758444
#> 11 0.0011884528
#> 12 0.0006456853
#> 13 0.0000000000
#> 14 0.0018579667
#> 15 0.0012555981
#> 16 0.0135895642
#> 17 0.1135926887
#> 18 0.1796737806
#> 19 0.4367495288
#> 20 0.9062670353
#> 21 0.9261173518
#> 22 0.8978655393
#> 23 0.8988121516
#> 24 0.8622608681
#> 25 0.4294018327
#> 26 0.2608309742
#> 27 0.4127799846
#> 28 0.2087552197
#> 29 0.2625767574
#> 30 0.2251831634
#> 31 0.2927117476
#> 32 0.2750335428
#> 33 0.2348241091
#> 34 0.2916726914
#> 35 0.2898663701
#> 36 0.3606723956
#> 37 1.0000000000
#> 38 1.0000000000
#> 39 0.8996866268