Display details on the penalization

An S4 method showing information on the penalization process, including the employed penalty functions and the model matrices they affect. Additionally, it reports the optimal values of the tuning parameters and the tuning parameter selection strategy. If the automatic procedure was used, the output would also show the value of the influence factor, and the number of two-steps iterations.

# S4 method for penfaPenalty
show(object)

Arguments

object
An object of class penfaPenalty, found in the Penalize slot from an object of penfa class.

See also

Examples

## --- Continuing the Examples from penfa manual page:
require(utils)
example("penfa", echo = FALSE)

#> Computing weights for alasso (ML estimates)... done.
#> 
#> Automatic procedure: 
#> Iteration  1 : 0.00298271 
#> Iteration  2 : 0.00452604 
#> 
#> Largest absolute gradient value: 12.76355181
#> Fisher information matrix is positive definite
#> Eigenvalue range: [180.2917, 9189645]
#> Trust region iterations: 15 
#> Factor solution: admissible 
#> Effective degrees of freedom: 27.12936  
#> 
#> Largest absolute gradient value: 0.00108292
#> Fisher information matrix is positive definite
#> Eigenvalue range: [2.904073, 76169.99]
#> Trust region iterations: 22 
#> Factor solution: admissible 
#> Computing VCOV ... done.
#> Effective degrees of freedom: 94  
#> 
#> Automatic procedure: 
#> Iteration  1 : 0.00215339 3061215.58594407 0.00275238 
#> Iteration  2 : 0.00196371 4134.86838014 0.00811975 
#> 
#> Largest absolute gradient value: 68.84019108
#> Fisher information matrix is positive definite
#> Eigenvalue range: [53.19646, 1.21016e+12]
#> Trust region iterations: 22 
#> Factor solution: admissible 
#> Effective degrees of freedom: 56.77369  

alasso_fit@Penalize

#>   Strategy                                                 auto
#>   Influence factor                                            4
#>   Number of two-steps (automatic)                             2
#>                                                                
#>   Penalty function:                                            
#>     Sparsity                                             alasso
#>                                                                
#>   Additional tuning parameter                                  
#>     alasso                                                    1
#>                                                                
#>   Optimal tuning parameter:                                    
#>     Sparsity                                                   
#>      - Factor loadings                                    0.005
#>                                                                

alasso_fitMG@Penalize

#>   Strategy                                                 auto
#>   Influence factor                                            4
#>   Number of two-steps (automatic)                             2
#>                                                                
#>   Penalty functions:                                           
#>     Sparsity                                             alasso
#>     Invariance                                           alasso
#>                                                                
#>   Additional tuning parameter                                  
#>     alasso                                                    1
#>                                                                
#>   Optimal tuning parameters:                                   
#>     Sparsity                                                   
#>      - Factor loadings                                    0.002
#>     Invariance                                                 
#>      - Factor loadings                                 4134.868
#>      - Intercepts                                         0.008
#>