Model-implied moments for a penfa object

An S4 method returning the model-implied moments for an object of class penfa. For every group, a list with the model-implied moments is returned: cov contains the implied covariance matrix, and mean the implied mean vector. If just the covariance matrix is analyzed, only the cov argument is returned.

# S4 method for penfa
fitted(object, labels = TRUE)

Arguments

object
An object of class penfa, found as a result of a call to penfa.

labels
Logical. If TRUE, the model-implied moments are named according to the item names used in the model syntax.

See also

Examples

## --- Continuing the Examples from penfa manual page:
require(utils)
example("penfa", echo = FALSE)

#> Computing weights for alasso (ML estimates)... done.
#> 
#> Automatic procedure: 
#> Iteration  1 : 0.00298271 
#> Iteration  2 : 0.00452604 
#> 
#> Largest absolute gradient value: 12.76355181
#> Fisher information matrix is positive definite
#> Eigenvalue range: [180.2917, 9189645]
#> Trust region iterations: 15 
#> Factor solution: admissible 
#> Effective degrees of freedom: 27.12936  
#> 
#> Largest absolute gradient value: 0.00108292
#> Fisher information matrix is positive definite
#> Eigenvalue range: [2.904073, 76169.99]
#> Trust region iterations: 22 
#> Factor solution: admissible 
#> Computing VCOV ... done.
#> Effective degrees of freedom: 94  
#> 
#> Automatic procedure: 
#> Iteration  1 : 0.00215339 3061215.58594407 0.00275238 
#> Iteration  2 : 0.00196371 4134.86838014 0.00811975 
#> 
#> Largest absolute gradient value: 68.84019108
#> Fisher information matrix is positive definite
#> Eigenvalue range: [53.19646, 1.21016e+12]
#> Trust region iterations: 22 
#> Factor solution: admissible 
#> Effective degrees of freedom: 56.77369  

fitted(alasso_fit)

#> $cov
#>    h1    h2    h3    h4    h5    h6    h7    v1    v2    v3    v4    v5   
#> h1 0.974                                                                  
#> h2 0.657 0.969                                                            
#> h3 0.594 0.665 0.973                                                      
#> h4 0.678 0.759 0.686 0.968                                                
#> h5 0.655 0.734 0.664 0.757 0.968                                          
#> h6 0.669 0.749 0.677 0.772 0.749 0.966                                    
#> h7 0.630 0.706 0.638 0.724 0.708 0.727 0.967                              
#> v1 0.571 0.640 0.579 0.652 0.649 0.672 0.681 0.969                        
#> v2 0.585 0.656 0.593 0.668 0.665 0.689 0.697 0.741 0.968                  
#> v3 0.566 0.634 0.573 0.645 0.642 0.665 0.674 0.716 0.734 0.970            
#> v4 0.566 0.634 0.573 0.646 0.643 0.666 0.674 0.717 0.735 0.710 0.970      
#> v5 0.541 0.606 0.547 0.617 0.614 0.636 0.644 0.685 0.701 0.678 0.678 0.972
#> 

fitted(alasso_fitMG)

#> $LEB
#> $LEB$cov
#>    h1    h2    h3    h4    h5    h6    h7    v1    v2    v3    v4    v5   
#> h1 1.135                                                                  
#> h2 0.831 1.131                                                            
#> h3 0.750 0.828 1.091                                                      
#> h4 0.872 0.963 0.869 1.137                                                
#> h5 0.833 0.919 0.829 0.964 1.083                                          
#> h6 0.850 0.938 0.847 0.984 0.940 1.124                                    
#> h7 0.813 0.898 0.810 0.941 0.899 0.920 1.105                              
#> v1 0.759 0.838 0.756 0.874 0.839 0.863 0.869 1.088                        
#> v2 0.771 0.851 0.768 0.888 0.852 0.877 0.882 0.925 1.146                  
#> v3 0.742 0.819 0.739 0.854 0.820 0.844 0.849 0.890 0.904 1.120            
#> v4 0.741 0.818 0.739 0.854 0.820 0.844 0.849 0.890 0.904 0.870 1.128      
#> v5 0.707 0.781 0.705 0.815 0.782 0.805 0.810 0.849 0.862 0.830 0.830 1.081
#> 
#> $LEB$mean
#>    h1    h2    h3    h4    h5    h6    h7    v1    v2    v3    v4    v5 
#> 0.092 0.088 0.092 0.047 0.127 0.110 0.114 0.148 0.127 0.144 0.144 0.143 
#> 
#> 
#> $TAIW
#> $TAIW$cov
#>    h1    h2    h3    h4    h5    h6    h7    v1    v2    v3    v4    v5   
#> h1 0.868                                                                  
#> h2 0.509 0.828                                                            
#> h3 0.459 0.507 0.869                                                      
#> h4 0.535 0.591 0.533 0.862                                                
#> h5 0.510 0.563 0.508 0.592 0.867                                          
#> h6 0.519 0.573 0.517 0.602 0.574 0.821                                    
#> h7 0.486 0.536 0.484 0.562 0.537 0.549 0.851                              
#> v1 0.428 0.472 0.426 0.491 0.473 0.489 0.508 0.887                        
#> v2 0.434 0.480 0.433 0.499 0.480 0.497 0.516 0.577 0.794                  
#> v3 0.418 0.462 0.417 0.480 0.462 0.478 0.497 0.555 0.564 0.815            
#> v4 0.418 0.461 0.416 0.480 0.462 0.478 0.497 0.555 0.564 0.543 0.803      
#> v5 0.399 0.440 0.397 0.458 0.441 0.456 0.474 0.530 0.538 0.518 0.517 0.854
#> 
#> $TAIW$mean
#>     h1     h2     h3     h4     h5     h6     h7     v1     v2     v3     v4 
#> -0.101 -0.102 -0.102 -0.057 -0.127 -0.115 -0.134 -0.162 -0.140 -0.159 -0.159 
#>     v5 
#> -0.150 
#> 
#>