S4 Class for describing a penfa model

The penfa class represents a (fitted) penalized factor analysis model. It contains a description of the model as specified by the user, a summary of the data, an internal matrix representation, the fitting results, and the penalized quantities.

Objects from the Class

Objects can be created via the penfa function.

Slots

version:

The penfa package version used to create this object.

call:

The function call as returned by match.call().

timing:

The elapsed time (user + system) for various parts of the program as a list, including the total time.

Options:

Named list of options that were provided by the user or filled-in automatically. See penfaOptions for additional details.

ParTable:

Named list describing the model parameters. Can be coerced to a data.frame. This is also called "parameter table". It includes information on the fixed, free and penalized parameters, their indices, the active penalization strategies ("none", "shrink", "diff", or "shrink + diff"), the starting values, the estimated parameters and the associated standard errors.

pta:

Named list containing parameter table attributes, like observed and latent variable names, their indices, and the number of groups.

Data:

Object of internal class "penfaData"; contains information about the data set. See the penfaData class for additional details.

SampleStats:

Object of internal class "penfaSampleStats"; contains the sample statistics. See the penfaSampleStats class for additional details.

Model:

Object of internal class "penfaModel": the internal (matrix) representation of the model. See the penfaModel class for additional details.

Optim:

List. Information about the optimization process. This includes the estimated parameters (x), the number of estimated parameters (npar), the number of trust-region iterations (iterations), the value of the penalized objective function (fx.pen), the value of the unpenalized objective function (fx.unpen), the penalized log-likelihood (logl.pen; this is equal to fx.pen multiplied by (-1)), the unpenalized log-likelihood (logl.unpen; this is equal to fx.unpen multiplied by (-1)), the penalized gradient (dx.pen), the penalized Hessian/Fisher information matrix (hessian.pen), the list of control arguments for the trust-region algorithm (control), and how many times the objective function became non-positive definite during the estimation process (npd). If penfa was called with the option verbose = TRUE, the following additional arguments coming from the trust-region function trust are reported in the Optim slot: argpath, argtry, type, accept, radii, rho, fx.val, fx.valtry, change, stepnorm. See the manual page of trust from the trust package for an overview of these quantities.

Penalize:

Object of internal class "penfaPenalty"; contains information about the penalization. See the penfaPenalty for additional details.

Implied:

List. Model-implied moments (covariance matrix and mean vector).

Vcov:

List. Information about the covariance matrix (vcov) of the model parameters. This slot includes the following quantities: the type of penalized information matrix used in the model (either Hessian or Fisher; information), the vcov matrix of parameters (vcov), whether the convergence checks on the penalized gradient and the penalized information matrix were satisfied (solution), whether the employed information matrix was positive-definite (pdef), whether the estimated factor solution was admissible (admissibility), the standard errors computed according to the Bayesian result from the information matrix reported in information (se), and the 95% confidence intervals (ci).

Inference:

List. Information on effective degrees of the model and information criteria for model selection. This slot reports the following quantities: effective degree of freedom for each parameter (edf.single), total edf (edf), influence matrix (influence.mat), generalized information criteria (IC), such as AIC and BIC.

external:

List. Empty slot.

Methods

The following methods are available for an object of class penfa:

show

signature(object = "penfa"): Prints a short summary of the estimation process, including the optimization method, the specified penalty functions, the convergence status, the number of iterations, the tuning selection strategy, and the effective degrees of freedom. See the manual page of show,penfa-method for details.

summary

signature(object = "penfa", header = TRUE, estimates = TRUE, ci = TRUE, level = 0.95, nd = 3L, cutoff = 0.05, extra = TRUE): Prints a summary of the model parameter estimates, and the optimization process. See the manual page of summary,penfa-method for details.

coef

signature(object = "penfa", type = "free", labels = TRUE): Returns the estimates of the parameters in the model as a named numeric vector. See the manual page of coef,penfa-method for details.

fitted

signature(object = "penfa", labels = TRUE): Returns a list of the model-implied moments (per group). See the manual page of fitted,penfa-method for details.

References

Geminiani, E., Marra, G., & Moustaki, I. (2021). "Single- and Multiple-Group Penalized Factor Analysis: A Trust-Region Algorithm Approach with Integrated Automatic Multiple Tuning Parameter Selection." Psychometrika, 86(1), 65-95. https://doi.org/10.1007/s11336-021-09751-8

See also

Examples

data(ccdata)

syntax = 'help  =~   h1 + h2 + h3 + h4 + h5 + h6 + h7 + 0*v1 + v2 + v3 + v4 + v5
          voice =~ 0*h1 + h2 + h3 + h4 + h5 + h6 + h7 +   v1 + v2 + v3 + v4 + v5'

alasso_fit <- penfa(## factor model
                    model  = syntax,
                    data   = ccdata,
                    std.lv = TRUE,
                    ## penalization
                    pen.shrink = "alasso",
                    eta = list(shrink = c("lambda" = 0.01), diff = c("none" = 0)),
                    ## automatic procedure
                    strategy = "auto")

#> Computing weights for alasso (ML estimates)... done.
#> 
#> Automatic procedure: 
#> Iteration  1 : 0.00298271 
#> Iteration  2 : 0.00452604 
#> 
#> Largest absolute gradient value: 12.76355181
#> Fisher information matrix is positive definite
#> Eigenvalue range: [180.2917, 9189645]
#> Trust region iterations: 15 
#> Factor solution: admissible 
#> Effective degrees of freedom: 27.12936  

alasso_fit

#> penfa 0.0.0.9000 reached convergence
#> 
#>   Number of observations                                    767
#>                                                                
#>   Estimator                                                PMLE
#>   Optimization method                              trust-region
#>   Information                                            fisher
#>   Strategy                                                 auto
#>   Number of iterations (total)                               58
#>   Number of two-steps (automatic)                             2
#>   Effective degrees of freedom                           27.129
#>                                                                
#>   Penalty function:                                            
#>     Sparsity                                             alasso
#>                                                                
#>                                                                

class(alasso_fit)

#> [1] "penfa"
#> attr(,"package")
#> [1] "penfa"

summary(alasso_fit)

#> penfa 0.0.0.9000 reached convergence
#> 
#>   Number of observations                                    767
#>   Number of groups                                            1
#>   Number of observed variables                               12
#>   Number of latent factors                                    2
#>                                                                
#>   Estimator                                                PMLE
#>   Optimization method                              trust-region
#>   Information                                            fisher
#>   Strategy                                                 auto
#>   Number of iterations (total)                               58
#>   Number of two-steps (automatic)                             2
#>   Influence factor                                            4
#>   Number of parameters:                                        
#>     Free                                                     13
#>     Penalized                                                22
#>   Effective degrees of freedom                           27.129
#>   GIC                                                 17222.980
#>   GBIC                                                17348.928
#>                                                                
#>   Penalty function:                                            
#>     Sparsity                                             alasso
#>                                                                
#>   Additional tuning parameter                                  
#>     alasso                                                    1
#>                                                                
#>   Optimal tuning parameter:                                    
#>     Sparsity                                                   
#>      - Factor loadings                                    0.005
#>                                                                
#> 
#> Parameter Estimates:
#> 
#> Latent Variables:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>   help =~                                                      
#>     h1               pen       0.766    0.030    0.707    0.825
#>     h2               pen       0.858    0.028    0.803    0.913
#>     h3               pen       0.775    0.030    0.717    0.834
#>     h4               pen       0.921    0.038    0.847    0.995
#>     h5               pen       0.810    0.040    0.732    0.887
#>     h6               pen       0.782    0.044    0.696    0.868
#>     h7               pen       0.523    0.050    0.426    0.620
#>     v1             fixed       0.000             0.000    0.000
#>     v2               pen       0.000                           
#>     v3               pen       0.000                           
#>     v4               pen       0.000                           
#>     v5               pen      -0.000                           
#>   voice =~                                                     
#>     h1             fixed       0.000             0.000    0.000
#>     h2               pen      -0.000                           
#>     h3               pen       0.000                           
#>     h4               pen      -0.041                           
#>     h5               pen       0.053    0.031   -0.008    0.114
#>     h6               pen       0.104    0.038    0.029    0.180
#>     h7               pen       0.341    0.049    0.246    0.437
#>     v1               pen       0.851    0.028    0.795    0.906
#>     v2               pen       0.871    0.028    0.817    0.926
#>     v3               pen       0.842    0.029    0.786    0.898
#>     v4               pen       0.843    0.029    0.787    0.899
#>     v5               pen       0.805    0.029    0.747    0.862
#> 
#> Covariances:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>   help ~~                                                      
#>     voice           free       0.877    0.011    0.855    0.900
#> 
#> Variances:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>    .h1              free       0.388    0.021    0.346    0.429
#>    .h2              free       0.233    0.014    0.205    0.261
#>    .h3              free       0.372    0.021    0.332    0.413
#>    .h4              free       0.184    0.012    0.160    0.209
#>    .h5              free       0.235    0.014    0.207    0.263
#>    .h6              free       0.201    0.012    0.177    0.225
#>    .h7              free       0.264    0.015    0.235    0.293
#>    .v1              free       0.245    0.015    0.216    0.275
#>    .v2              free       0.208    0.014    0.182    0.235
#>    .v3              free       0.261    0.016    0.230    0.292
#>    .v4              free       0.259    0.016    0.228    0.290
#>    .v5              free       0.324    0.019    0.287    0.361
#>     help           fixed       1.000             1.000    1.000
#>     voice          fixed       1.000             1.000    1.000
#> 

coef(alasso_fit)

#>    help=~h1    help=~h2    help=~h3    help=~h4    help=~h5    help=~h6 
#>       0.766       0.858       0.775       0.921       0.810       0.782 
#>    help=~h7    help=~v2    help=~v3    help=~v4    help=~v5   voice=~h2 
#>       0.523       0.000       0.000       0.000       0.000       0.000 
#>   voice=~h3   voice=~h4   voice=~h5   voice=~h6   voice=~h7   voice=~v1 
#>       0.000      -0.041       0.053       0.104       0.341       0.851 
#>   voice=~v2   voice=~v3   voice=~v4   voice=~v5      h1~~h1      h2~~h2 
#>       0.871       0.842       0.843       0.805       0.388       0.233 
#>      h3~~h3      h4~~h4      h5~~h5      h6~~h6      h7~~h7      v1~~v1 
#>       0.372       0.184       0.235       0.201       0.264       0.245 
#>      v2~~v2      v3~~v3      v4~~v4      v5~~v5 help~~voice 
#>       0.208       0.261       0.259       0.324       0.877 

fitted(alasso_fit)

#> $cov
#>    h1    h2    h3    h4    h5    h6    h7    v1    v2    v3    v4    v5   
#> h1 0.974                                                                  
#> h2 0.657 0.969                                                            
#> h3 0.594 0.665 0.973                                                      
#> h4 0.678 0.759 0.686 0.968                                                
#> h5 0.655 0.734 0.664 0.757 0.968                                          
#> h6 0.669 0.749 0.677 0.772 0.749 0.966                                    
#> h7 0.630 0.706 0.638 0.724 0.708 0.727 0.967                              
#> v1 0.571 0.640 0.579 0.652 0.649 0.672 0.681 0.969                        
#> v2 0.585 0.656 0.593 0.668 0.665 0.689 0.697 0.741 0.968                  
#> v3 0.566 0.634 0.573 0.645 0.642 0.665 0.674 0.716 0.734 0.970            
#> v4 0.566 0.634 0.573 0.646 0.643 0.666 0.674 0.717 0.735 0.710 0.970      
#> v5 0.541 0.606 0.547 0.617 0.614 0.636 0.644 0.685 0.701 0.678 0.678 0.972
#>