Summary constructor for a penfa object

An S4 method printing a summary of the model parameter estimates for an object of class penfa.

# S4 method for penfa
summary(
  object,
  header = TRUE,
  estimates = TRUE,
  ci = TRUE,
  level = 0.95,
  nd = 3L,
  cutoff = 0.05,
  extra = TRUE
)

Arguments

object
An object of class penfa, found as a result of a call to penfa.

header
Logical. If TRUE, the header section is printed. The header contains relevant information about the data, the fitted model, the optimization process, and the penalization strategy, including, for instance, the employed penalties, the estimated effective degrees of freedom (edf), the optimal values of the tuning parameter(s), the GBIC and many others.

estimates
Logical. If TRUE, a section with the parameter estimates is printed out.

ci
Logical. If TRUE, confidence intervals are added to the parameter estimates section.

level
Logical. It denotes the significance level used for the statistical tests.

nd
Integer. It determines the number of digits after the decimal point to be printed in the parameter estimates section.

cutoff
Numeric. Standard errors and confidence intervals for the penalized parameter estimates falling below the cutoff value are not displayed. Confidence intervals for the parameters that have been penalized and shrunken to zero must be treated with caution.

extra
Logical. If TRUE, additional information on the model are displayed.

See also

Examples

## --- Continuing the Examples from penfa manual page:
require(utils)
example("penfa", echo = FALSE)

#> Computing weights for alasso (ML estimates)... done.
#> 
#> Automatic procedure: 
#> Iteration  1 : 0.00298271 
#> Iteration  2 : 0.00452604 
#> 
#> Largest absolute gradient value: 12.76355181
#> Fisher information matrix is positive definite
#> Eigenvalue range: [180.2917, 9189645]
#> Trust region iterations: 15 
#> Factor solution: admissible 
#> Effective degrees of freedom: 27.12936  
#> 
#> Largest absolute gradient value: 0.00108292
#> Fisher information matrix is positive definite
#> Eigenvalue range: [2.904073, 76169.99]
#> Trust region iterations: 22 
#> Factor solution: admissible 
#> Computing VCOV ... done.
#> Effective degrees of freedom: 94  
#> 
#> Automatic procedure: 
#> Iteration  1 : 0.00215674 6115.68152854 0.00300201 
#> Iteration  2 : 0.00196778 4011.40169767 0.00772393 
#> 
#> Largest absolute gradient value: 61.89552041
#> Fisher information matrix is positive definite
#> Eigenvalue range: [53.00923, 1.174025e+12]
#> Trust region iterations: 18 
#> Factor solution: admissible 
#> Effective degrees of freedom: 56.75744  

summary(alasso_fit)

#> penfa 0.0.0.9000 reached convergence
#> 
#>   Number of observations                                    767
#>   Number of groups                                            1
#>   Number of observed variables                               12
#>   Number of latent factors                                    2
#>                                                                
#>   Estimator                                                PMLE
#>   Optimization method                              trust-region
#>   Information                                            fisher
#>   Strategy                                                 auto
#>   Number of iterations (total)                               58
#>   Number of two-steps (automatic)                             2
#>   Influence factor                                            4
#>   Number of parameters:                                        
#>     Free                                                     13
#>     Penalized                                                22
#>   Effective degrees of freedom                           27.129
#>   GIC                                                 17222.980
#>   GBIC                                                17348.928
#>                                                                
#>   Penalty function:                                            
#>     Sparsity                                             alasso
#>                                                                
#>   Additional tuning parameter                                  
#>     alasso                                                    1
#>                                                                
#>   Optimal tuning parameter:                                    
#>     Sparsity                                                   
#>      - Factor loadings                                    0.005
#>                                                                
#> 
#> Parameter Estimates:
#> 
#> Latent Variables:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>   help =~                                                      
#>     h1               pen       0.766    0.030    0.707    0.825
#>     h2               pen       0.858    0.028    0.803    0.913
#>     h3               pen       0.775    0.030    0.717    0.834
#>     h4               pen       0.921    0.038    0.847    0.995
#>     h5               pen       0.810    0.040    0.732    0.887
#>     h6               pen       0.782    0.044    0.696    0.868
#>     h7               pen       0.523    0.050    0.426    0.620
#>     v1             fixed       0.000             0.000    0.000
#>     v2               pen       0.000                           
#>     v3               pen       0.000                           
#>     v4               pen       0.000                           
#>     v5               pen      -0.000                           
#>   voice =~                                                     
#>     h1             fixed       0.000             0.000    0.000
#>     h2               pen      -0.000                           
#>     h3               pen       0.000                           
#>     h4               pen      -0.041                           
#>     h5               pen       0.053    0.031   -0.008    0.114
#>     h6               pen       0.104    0.038    0.029    0.180
#>     h7               pen       0.341    0.049    0.246    0.437
#>     v1               pen       0.851    0.028    0.795    0.906
#>     v2               pen       0.871    0.028    0.817    0.926
#>     v3               pen       0.842    0.029    0.786    0.898
#>     v4               pen       0.843    0.029    0.787    0.899
#>     v5               pen       0.805    0.029    0.747    0.862
#> 
#> Covariances:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>   help ~~                                                      
#>     voice           free       0.877    0.011    0.855    0.900
#> 
#> Variances:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>    .h1              free       0.388    0.021    0.346    0.429
#>    .h2              free       0.233    0.014    0.205    0.261
#>    .h3              free       0.372    0.021    0.332    0.413
#>    .h4              free       0.184    0.012    0.160    0.209
#>    .h5              free       0.235    0.014    0.207    0.263
#>    .h6              free       0.201    0.012    0.177    0.225
#>    .h7              free       0.264    0.015    0.235    0.293
#>    .v1              free       0.245    0.015    0.216    0.275
#>    .v2              free       0.208    0.014    0.182    0.235
#>    .v3              free       0.261    0.016    0.230    0.292
#>    .v4              free       0.259    0.016    0.228    0.290
#>    .v5              free       0.324    0.019    0.287    0.361
#>     help           fixed       1.000             1.000    1.000
#>     voice          fixed       1.000             1.000    1.000
#> 

summary(alasso_fitMG)

#> penfa 0.0.0.9000 reached convergence
#> 
#>   Number of observations per group:                            
#>     LEB                                                     402
#>     TAIW                                                    365
#>   Number of groups                                            2
#>   Number of observed variables                               12
#>   Number of latent factors                                    2
#>                                                                
#>   Estimator                                                PMLE
#>   Optimization method                              trust-region
#>   Information                                            fisher
#>   Strategy                                                 auto
#>   Number of iterations (total)                              150
#>   Number of two-steps (automatic)                             2
#>   Influence factor                                            4
#>   Number of parameters:                                        
#>     Free                                                     34
#>     Penalized                                                60
#>   Effective degrees of freedom                           56.757
#>   GIC                                                 17096.169
#>   GBIC                                                17359.665
#>                                                                
#>   Penalty functions:                                           
#>     Sparsity                                             alasso
#>     Invariance                                           alasso
#>                                                                
#>   Additional tuning parameter                                  
#>     alasso                                                    1
#>                                                                
#>   Optimal tuning parameters:                                   
#>     Sparsity                                                   
#>      - Factor loadings                                    0.002
#>     Invariance                                                 
#>      - Factor loadings                                 4011.402
#>      - Intercepts                                         0.008
#>                                                                
#> 
#> Parameter Estimates:
#> 
#> 
#> Group 1 [LEB]:
#> 
#> Latent Variables:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>   help =~                                                      
#>     h1             fixed       1.000             1.000    1.000
#>     h2               pen       1.103    0.039    1.027    1.180
#>     h3               pen       0.996    0.040    0.917    1.076
#>     h4               pen       1.196    0.050    1.099    1.294
#>     h5               pen       1.106    0.038    1.031    1.181
#>     h6               pen       1.084    0.051    0.984    1.185
#>     h7               pen       0.742    0.065    0.614    0.870
#>     v1             fixed       0.000             0.000    0.000
#>     v2               pen       0.000                           
#>     v3               pen       0.000                           
#>     v4               pen       0.000                           
#>     v5               pen      -0.000                           
#>   voice =~                                                     
#>     h1             fixed       0.000             0.000    0.000
#>     h2               pen       0.000                           
#>     h3               pen      -0.000                           
#>     h4               pen      -0.038                           
#>     h5               pen       0.000                           
#>     h6               pen       0.044                           
#>     h7               pen       0.336    0.056    0.225    0.446
#>     v1             fixed       1.000             1.000    1.000
#>     v2               pen       1.015    0.028    0.960    1.070
#>     v3               pen       0.977    0.029    0.920    1.034
#>     v4               pen       0.977    0.029    0.920    1.034
#>     v5               pen       0.932    0.030    0.873    0.991
#> 
#> Covariances:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>   help ~~                                                      
#>     voice           free       0.759    0.062    0.637    0.882
#> 
#> Intercepts:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>    .h1             fixed       0.000             0.000    0.000
#>    .v1             fixed       0.000             0.000    0.000
#>    .h2               pen      -0.012    0.031   -0.073    0.049
#>    .h3               pen      -0.001    0.032   -0.062    0.061
#>    .h4               pen      -0.058    0.033   -0.122    0.007
#>    .h5               pen       0.024    0.031   -0.037    0.084
#>    .h6               pen       0.003    0.029   -0.055    0.060
#>    .h7               pen      -0.004    0.025   -0.054    0.045
#>    .v2               pen      -0.022    0.026   -0.072    0.029
#>    .v3               pen      -0.001    0.025   -0.050    0.048
#>    .v4               pen      -0.001    0.025   -0.050    0.048
#>    .v5               pen       0.004    0.026   -0.047    0.055
#>     help            free       0.092    0.049   -0.005    0.189
#>     voice           free       0.148    0.051    0.048    0.248
#> 
#> Variances:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>    .h1              free       0.382    0.029    0.326    0.438
#>    .h2              free       0.213    0.017    0.180    0.247
#>    .h3              free       0.344    0.026    0.293    0.395
#>    .h4              free       0.126    0.012    0.103    0.150
#>    .h5              free       0.162    0.014    0.135    0.189
#>    .h6              free       0.164    0.014    0.137    0.191
#>    .h7              free       0.210    0.016    0.178    0.241
#>    .v1              free       0.176    0.016    0.145    0.208
#>    .v2              free       0.207    0.018    0.171    0.242
#>    .v3              free       0.250    0.021    0.209    0.290
#>    .v4              free       0.259    0.021    0.217    0.300
#>    .v5              free       0.290    0.023    0.245    0.335
#>     help            free       0.753    0.069    0.617    0.889
#>     voice           free       0.911    0.074    0.766    1.056
#> 
#> 
#> Group 2 [TAIW]:
#> 
#> Latent Variables:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>   help =~                                                      
#>     h1             fixed       1.000             1.000    1.000
#>     h2               pen       1.103    0.039    1.027    1.180
#>     h3               pen       0.996    0.040    0.917    1.076
#>     h4               pen       1.196    0.050    1.099    1.294
#>     h5               pen       1.106    0.038    1.031    1.181
#>     h6               pen       1.084    0.051    0.984    1.185
#>     h7               pen       0.742    0.065    0.614    0.870
#>     v1             fixed       0.000             0.000    0.000
#>     v2               pen       0.000                           
#>     v3               pen       0.000                           
#>     v4               pen       0.000                           
#>     v5               pen      -0.000                           
#>   voice =~                                                     
#>     h1             fixed       0.000             0.000    0.000
#>     h2               pen      -0.000                           
#>     h3               pen       0.000                           
#>     h4               pen      -0.038                           
#>     h5               pen       0.000                           
#>     h6               pen       0.044                           
#>     h7               pen       0.336    0.056    0.225    0.446
#>     v1             fixed       1.000             1.000    1.000
#>     v2               pen       1.015    0.028    0.960    1.070
#>     v3               pen       0.977    0.029    0.920    1.034
#>     v4               pen       0.977    0.029    0.920    1.034
#>     v5               pen       0.932    0.030    0.873    0.991
#> 
#> Covariances:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>   help ~~                                                      
#>     voice           free       0.428    0.039    0.351    0.505
#> 
#> Intercepts:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>    .h1             fixed       0.000             0.000    0.000
#>    .v1             fixed       0.000             0.000    0.000
#>    .h2               pen       0.008    0.032   -0.054    0.070
#>    .h3               pen      -0.001    0.032   -0.062    0.061
#>    .h4               pen       0.058    0.036   -0.013    0.129
#>    .h5               pen      -0.012    0.032   -0.075    0.051
#>    .h6               pen       0.003    0.029   -0.055    0.060
#>    .h7               pen      -0.004    0.025   -0.054    0.046
#>    .v2               pen       0.022    0.027   -0.030    0.075
#>    .v3               pen      -0.001    0.025   -0.050    0.048
#>    .v4               pen      -0.001    0.025   -0.050    0.048
#>    .v5               pen       0.002    0.026   -0.049    0.053
#>     help            free      -0.101    0.043   -0.186   -0.017
#>     voice           free      -0.161    0.045   -0.249   -0.074
#> 
#> Variances:
#>                     Type    Estimate  Std.Err     2.5%    97.5%
#>    .h1              free       0.407    0.033    0.343    0.471
#>    .h2              free       0.267    0.023    0.222    0.312
#>    .h3              free       0.411    0.033    0.346    0.476
#>    .h4              free       0.240    0.022    0.197    0.283
#>    .h5              free       0.303    0.026    0.253    0.354
#>    .h6              free       0.237    0.021    0.196    0.278
#>    .h7              free       0.321    0.026    0.270    0.371
#>    .v1              free       0.318    0.027    0.265    0.372
#>    .v2              free       0.208    0.020    0.169    0.247
#>    .v3              free       0.272    0.024    0.225    0.319
#>    .v4              free       0.261    0.023    0.215    0.306
#>    .v5              free       0.361    0.030    0.302    0.419
#>     help            free       0.461    0.045    0.373    0.549
#>     voice           free       0.568    0.051    0.469    0.668
#>